# Speedup byte/hex/number conversions