* 55% faster bytesToNumber
* 40% faster bytesToHex
* 30% faster hexToBytes
* 25% numTo32b