@paulmillr 2022-03-15
1. `setBiguint64` is not available in safari 14, we can't use it. Unless we shim like in **noble-hashes**
2. Could you post some online benchmark for `bytesToHex`?
3. Could you post results of `npm run bench` before/after improved `bytesToNumber` (only 1 function)?