@brandonblack 2022-03-15
1. Ah, I thought I could just target ES2020.

Will do some more benchmarking and post results. I didn't benchmark in browsers, only on node.