@brandonblack 2022-03-15
You are absolutely right on `bytesToHex` don't know how I messed up that result.

Here's a micro-bmark of these 2 functions:
```
Old:
bytesToNumber x 561,797 ops/sec @ 1μs/op
bytesToHex x 1,733,102 ops/sec @ 577ns/op

New:
bytesToNumber x 935,453 ops/sec @ 1μs/op
bytesToHex x 750,750 ops/sec @ 1μs/op
```