@brandonblack 2022-03-15
Without `getBigUint64` this `bytesToNumber` is about the same as the one in `main`:
```
bytesToNumber x 590,667 ops/sec @ 1μs/op
```