@paulmillr 2022-04-04
closing since the actual speed improvements are tiny compared to complexity